+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Adjust Prisma import on error-encoder
 - Adjust body upload size from 500mb to 900mb
 
+### Fixed
+
+- Fixed change password error deleting password-token-verification
+
 ## 1.0.0
 
 ### Added